move regress test to table package
