+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-* Fix "Unexpected NullFlag while parse Optional nil JSONDocument value"
+* Fixed Unexpected NullFlag while parse nil JSONDocument value
 * Removed `internal/conn/conn.streamUsages` and `internal/conn/conn.usages` (`internal/conn.conn` always touching last usage timestamp on API calls)
 * Removed auto-reconnecting for broken conns
 * Refactored applying actual endpoints list after re-discovery (replaced diff-merge logic to swap cluster struct, cluster and balancers are immutable now)
